#40 update file structure, update webpack config
