+++ src/entries.js
@@ -1,0 +1,5 @@
+'use strict';
+
+module.exports = {
+  index: './pages/index/index.js',
+};
