+++ src/pages/index/index.html
@@ -1,4 +1,4 @@
-﻿<!DOCTYPE html>
+﻿<!doctype html>
 <html lang="en-us" dir="ltr" xmlns="http://www.w3.org/1999/xhtml">
   <head>
     <title>Eugene Serb – Tetris</title>
